KKJJ-57 <refactor> (<Popup>) : 팝업 이미지 드라이브에 추가 및 엔티티 경로 이동
