--- src/main/java/renewal/awesome_travel_backoffice/popup/entity/Popup.java
@@ -1,44 +1,0 @@
-package renewal.awesome_travel_backoffice.popup.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import renewal.common.entity.AuditingFields;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@ToString
-@Table(name = "popup")
-public class Popup extends AuditingFields {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, name = "display_order")
-    private Integer displayOrder;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private LocalDate startDate;
-
-    @Column(nullable = false)
-    private LocalDate endDate;
-
-    @Column(nullable = false)
-    private boolean active = true;
-
-    @Column(nullable = false, length = 512)
-    private String file;
-
-    @Column(nullable = false, length = 512)
-    private String url;
-}
